Update CEMO005 and CEMO006
